requirements.txt
amqp==2.2.2
billiard==3.5.0.3
cachetools==2.0.1
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
google-api-python-client==1.6.4
google-auth==1.2.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
oauth2client==4.1.2
oauthlib==2.0.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
SQLAlchemy==1.1.15
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
Werkzeug==0.13
